backend/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
openai==1.54.0
httpx==0.27.2
pdfplumber==0.11.4
python-docx==1.1.2
pydantic==2.9.2
python-dotenv==1.0.1
supabase==2.9.0
docx2txt==0.8
